move Env struct to envvars
